move all modules to `/lib`
Follows up #4 moving the remaining modules into the `/lib` directory.
