+++ api/discord_app/sub/submit.ts
@@ -2,12 +2,12 @@ import type {
   APIApplicationCommandInteractionDataOption,
   APIApplicationCommandOption,
   APIInteractionResponse,
-} from "../../deps.ts";
+} from "../../../deps.ts";
 import {
   ApplicationCommandOptionType,
   InteractionResponseType,
-} from "../../deps.ts";
-import type * as leaderboard from "../../leaderboard/mod.ts";
+} from "../../../deps.ts";
+import type * as leaderboard from "../../../lib/leaderboard/mod.ts";
 
 export const SUBMIT = "submit";
 export const SUBMIT_DESCRIPTION =
