+++ lib/lc/fake_client.ts
